refactor(frontend): move formatters out of types/ + consolidate duplicates (#5107)
Deleted dead code (0 callers):
- types/batch-processing.formatRelativeTime, formatDateTime, formatDuration
- types/validation.formatTimestamp, formatRelativeTime, formatComponentName
- utils/workflowComponents.formatTimestamp, formatRelativeTime

Migrated runtime formatters out of types/ (antipattern):
- types/operations.formatRelativeTime -> utils/formatHelpers.formatTimeAgo (1 caller: OperationsList.vue)
- types/operations.formatDuration -> utils/formatHelpers.formatDuration (1 caller: OperationDetail.vue)
- types/audit.formatAuditTimestamp -> utils/formatHelpers.formatDateTime (1 caller: AuditLogTable.vue, body was byte-equivalent)
- types/audit.formatAuditRelativeTime -> utils/formatHelpers.formatTimeAgo (1 caller: AuditLogTable.vue)

Added to utils/formatHelpers.ts:
- formatDuration(startTime, endTime) -- shared helper for duration rendering

types/*.ts files now contain ONLY type declarations and non-format runtime
helpers (isTerminalStatus, canCancel, canResume, calculateSuccessRate,
getDateRangeForFilter, getHealthClass, canCancelJob, AUDIT_RESULT_CONFIG,
DEFAULT_AUDIT_FILTER, STATUS_CONFIG, etc.).

Verification:
- grep -rn 'export .* format' autobot-frontend/src/types/ returns nothing
- vue-tsc error count unchanged at 169 (matches Dev_new_gui baseline #5096)
- Zero errors in any modified file
- eslint: 0 errors, 2 pre-existing warnings

Closes #5107
