+++ autobot-frontend/src/types/batch-processing.ts
@@ -263,51 +263,3 @@ export function canCancelJob(job: BatchJob): boolean {
   return ['pending', 'running'].includes(job.status)
 }
 
-/**
- * Format duration in human-readable form
- */
-export function formatDuration(startTime: string | null | undefined, endTime: string | null | undefined): string {
-  if (!startTime) return '-'
-
-  const start = new Date(startTime).getTime()
-  const end = endTime ? new Date(endTime).getTime() : Date.now()
-  const durationMs = end - start
-
-  if (durationMs < 1000) return `${durationMs}ms`
-  if (durationMs < 60000) return `${Math.round(durationMs / 1000)}s`
-  if (durationMs < 3600000) {
-    const mins = Math.floor(durationMs / 60000)
-    const secs = Math.round((durationMs % 60000) / 1000)
-    return `${mins}m ${secs}s`
-  }
-
-  const hours = Math.floor(durationMs / 3600000)
-  const mins = Math.floor((durationMs % 3600000) / 60000)
-  return `${hours}h ${mins}m`
-}
-
-/**
- * Format timestamp to relative time
- */
-export function formatRelativeTime(timestamp: string | null | undefined): string {
-  if (!timestamp) return '-'
-
-  const date = new Date(timestamp)
-  const now = new Date()
-  const diffMs = now.getTime() - date.getTime()
-
-  if (diffMs < 60000) return 'Just now'
-  if (diffMs < 3600000) return `${Math.floor(diffMs / 60000)} min ago`
-  if (diffMs < 86400000) return `${Math.floor(diffMs / 3600000)} hours ago`
-  if (diffMs < 604800000) return `${Math.floor(diffMs / 86400000)} days ago`
-
-  return date.toLocaleDateString()
-}
-
-/**
- * Format datetime for display
- */
-export function formatDateTime(timestamp: string | null | undefined): string {
-  if (!timestamp) return '-'
-  return new Date(timestamp).toLocaleString()
-}
