+++ autobot-frontend/src/types/operations.ts
@@ -252,44 +252,3 @@ export function canCancel(operation: Operation): boolean {
 export function canResume(operation: Operation): boolean {
   return operation.can_resume && ['failed', 'timeout', 'paused'].includes(operation.status)
 }
-
-/**
- * Format duration in human-readable form
- */
-export function formatDuration(startTime: string | null, endTime: string | null): string {
-  if (!startTime) return '-'
-
-  const start = new Date(startTime).getTime()
-  const end = endTime ? new Date(endTime).getTime() : Date.now()
-  const durationMs = end - start
-
-  if (durationMs < 1000) return `${durationMs}ms`
-  if (durationMs < 60000) return `${Math.round(durationMs / 1000)}s`
-  if (durationMs < 3600000) {
-    const mins = Math.floor(durationMs / 60000)
-    const secs = Math.round((durationMs % 60000) / 1000)
-    return `${mins}m ${secs}s`
-  }
-
-  const hours = Math.floor(durationMs / 3600000)
-  const mins = Math.floor((durationMs % 3600000) / 60000)
-  return `${hours}h ${mins}m`
-}
-
-/**
- * Format timestamp to relative time
- */
-export function formatRelativeTime(timestamp: string | null): string {
-  if (!timestamp) return '-'
-
-  const date = new Date(timestamp)
-  const now = new Date()
-  const diffMs = now.getTime() - date.getTime()
-
-  if (diffMs < 60000) return 'Just now'
-  if (diffMs < 3600000) return `${Math.floor(diffMs / 60000)} min ago`
-  if (diffMs < 86400000) return `${Math.floor(diffMs / 3600000)} hours ago`
-  if (diffMs < 604800000) return `${Math.floor(diffMs / 86400000)} days ago`
-
-  return date.toLocaleDateString()
-}
