+++ autobot-frontend/src/types/validation.ts
@@ -114,43 +114,3 @@ export function getHealthClass(score: number): HealthClass {
   return 'critical'
 }
 
-/**
- * Format component name for display
- */
-export function formatComponentName(name: string): string {
-  return name
-    .replace(/_/g, ' ')
-    .replace(/\b\w/g, l => l.toUpperCase())
-}
-
-/**
- * Format timestamp for display
- */
-export function formatTimestamp(timestamp: string): string {
-  if (!timestamp) return 'N/A'
-  try {
-    return new Date(timestamp).toLocaleTimeString()
-  } catch {
-    return timestamp
-  }
-}
-
-/**
- * Format relative timestamp
- */
-export function formatRelativeTime(timestamp: string): string {
-  if (!timestamp) return 'N/A'
-  try {
-    const date = new Date(timestamp)
-    const now = new Date()
-    const diffMs = now.getTime() - date.getTime()
-    const diffMins = Math.floor(diffMs / 60000)
-
-    if (diffMins < 1) return 'Just now'
-    if (diffMins < 60) return `${diffMins}m ago`
-    if (diffMins < 1440) return `${Math.floor(diffMins / 60)}h ago`
-    return date.toLocaleDateString()
-  } catch {
-    return timestamp
-  }
-}
