+++ autobot-frontend/src/utils/workflowComponents.ts
@@ -92,45 +92,6 @@ export function getNotificationColor(type: NotificationType): string {
 }
 
 // ============================================================================
-// Time Utilities
-// ============================================================================
-
-/**
- * Format timestamp to locale time string
- *
- * @param timestamp - Unix timestamp in milliseconds
- * @returns Formatted time string (e.g., "2:30:45 PM")
- *
- * @example
- * ```ts
- * const time = formatTimestamp(Date.now()) // "2:30:45 PM"
- * ```
- */
-export function formatTimestamp(timestamp: number): string {
-  return new Date(timestamp).toLocaleTimeString()
-}
-
-/**
- * Format timestamp to relative time (e.g., "2 minutes ago")
- *
- * @param timestamp - Unix timestamp in milliseconds
- * @returns Relative time string
- */
-export function formatRelativeTime(timestamp: number): string {
-  const now = Date.now()
-  const diff = now - timestamp
-  const seconds = Math.floor(diff / 1000)
-  const minutes = Math.floor(seconds / 60)
-  const hours = Math.floor(minutes / 60)
-  const days = Math.floor(hours / 24)
-
-  if (seconds < 60) return 'just now'
-  if (minutes < 60) return `${minutes} minute${minutes > 1 ? 's' : ''} ago`
-  if (hours < 24) return `${hours} hour${hours > 1 ? 's' : ''} ago`
-  return `${days} day${days > 1 ? 's' : ''} ago`
-}
-
-// ============================================================================
 // Progress Utilities
 // ============================================================================
 
@@ -288,9 +249,6 @@ export default {
   // Icon utilities
   getNotificationIcon,
   getNotificationColor,
-  // Time utilities
-  formatTimestamp,
-  formatRelativeTime,
   // Progress utilities
   calculateProgress,
   getProgressColor,
